change hibernate pk as long

+++ src/main/java/com/bulain/hibernate/core/BasicMapper.java
@@ -1,8 +1,8 @@
 package com.bulain.hibernate.core;
 
 public interface BasicMapper<T> {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
     int insert(T record);
-    T selectByPrimaryKey(Integer id);
+    T selectByPrimaryKey(Long id);
     int updateByPrimaryKey(T record);
 }
