+++ src/main/java/com/bulain/hibernate/core/BasicService.java
@@ -3,9 +3,9 @@ package com.bulain.hibernate.core;
 import java.util.List;
 
 public interface BasicService<T> {
-    T get(Integer id, List<String> properties);
+    T get(Long id, List<String> properties);
     void insert(T record);
     void update(T record);
-    void delete(Integer id);
+    void delete(Long id);
     long countByDuplicate(T record);
 }
