+++ src/migration/scripts/20111024115340_create_table_persons.sql
@@ -1,7 +1,7 @@
 --// create table persons
 -- Migration SQL that makes the change goes here.
 create table persons(
-    id int(11) NOT NULL auto_increment,
+    id bigint(11) NOT NULL auto_increment,
     first_name varchar(20),
     last_name varchar(20),
